requirements.txt
requests>=2.22.0
flask>=1.1.1
pygogo>=0.12.0
python-dateutil>=2.8.1
meza>=0.41.1,<0.42.0
python-dotenv>=0.10.3
